Model updates
Shrank the Host model - now we have to fetch the rest of the host data with other user on the event detail
Now the host is one object, rather than a List<Host>
Updated Event and User models
